api/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
gunicorn==22.0.0
